Refactoring PropertyScaffold to widgets folder
