+++ lib/property_widgets/number_property.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import '../story.dart';
-import './properties.dart';
+import './widgets/property_scafold.dart';
 
 class NumberProperty extends StatefulWidget {
   final Property<double> property;
