+++ lib/property_widgets/text_property.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../story.dart';
-import './properties.dart';
+import './widgets/property_scafold.dart';
 
 class TextProperty extends StatefulWidget {
   final Property<String> property;
